Changed to using unsigned short ints for RGB values to save on memory. No speed up seen, but there is a space saving of 50 to 75 percent on the c code.

+++ blurry.cl
@@ -1,4 +1,4 @@
-__kernel void blurry(__global int* pixels, __global int* pixelsOut, int numRGBElements, int width, int height){
+__kernel void blurry(__global ushort* pixels, __global ushort* pixelsOut, int numRGBElements, int width, int height){
 
   pixelsOut[0] = 0;
   pixelsOut[1] = 0;
